Fixed 'Delete account by Id'.
Current state:
If a delete request is sent to /accounts/{id}, it will delete the
account row along with the corresponding user, admin or provider row.
If a delete request is sent to /users, /admins or /providers, it will
only work if the account has the corresponding role in which the request
is sent to.