Made the Header update when a user logs in.
Added an isLoggedIn component to the header to remove the text:
Login | Register, which is next to the user icon, when the user logs in.
Also changed the functionality behind the button after the user is
logged in. Will now redirect the user to the /account/account instead of
opening the login-screen if the user is logged in.