Integrated API fetches and removed static templates from MyRentals.
Removed the static object reference from App.jsx that was used as a template in MyRentals.
Implemented a fetch request in MyRentals to load car data from the /cars endpoint.
Updated MyRentalsCarDisplay to fetch rental data from the /rentals endpoint.
Connected the fetched data to the JSX rendering of MyRentals so that the page now displays dynamic data.