Fixed problem in MyRentalsCarDisplay.
Error was that if the user is fast enough after accessing the 'My Rentals' page, the user can try to access a RentalTable which hasn't been loaded yet. This will throw an error to the user. Fixed it by implementing a isLoading-useState. Will now display 'Loading rentals...' if accesses until the table-data has been loaded.