Implemented makeApiRequest in JwtUtility.js. It utilizes /refresh-token on expired tokens.
Also added an implementation of it in DeleteAccount (left the old implementation commented out for the sake of being able to be used as a reference for when refactoring the rest of the fetch requests. Added RootController and FallbackController classes. RootController holds the root / endpoint, will return the home page. FallbackController is for all unmatched routes.